refactor(components): migrate all Button usage to atomic Button component
- Remove legacy Button.jsx component to prevent conflicts
- Update Button.jsx API: remove content prop, use children instead
- Add height prop support for better component flexibility
- Migrate 7 files to use new atomic Button component:
  * Login.jsx, Auth/LoginFlow.jsx, Header/LoginModal.jsx
  * Register/FirstStep.jsx, Register/SecondStep.jsx, Register/ThirdStep.jsx
- Improve Button.styles.jsx with hover states and theme variants
- Ensure consistent button behavior across all components

Commit message organized by Claude Code assistant
