+++ src/container/Header/components/LoginModal.jsx
@@ -3,7 +3,7 @@ import Modal from '@components/Modal'
 import Account from '@container/Account'
 import styled from '@emotion/styled'
 import { loginBanner } from '@assets/images/index'
-import Button from '@components/Button'
+import Button from '@components/atoms/Button/Button'
 
 const LoginModal = ({
   isShow,
